pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pcrl"
version = "0.1.0"
description = "PPO/TRPO baselines with explicit hooks for predictive-coding RL research."
requires-python = ">=3.10"
readme = "README.md"
dependencies = [
    "torch>=2.1",
    "gymnasium[atari,accept-rom-license,mujoco]>=0.29",
    "numpy>=1.24",
    "pyyaml>=6.0",
    "opencv-python>=4.8",
    "tqdm>=4.65",
    "pandas>=2.0",
    "matplotlib>=3.7",
]

[project.optional-dependencies]
dev = ["pytest>=7", "ruff>=0.4"]

[tool.setuptools.packages.find]
include = ["pcrl*"]
